+++ health_check/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+"""
+Django HealthCheck application.
+"""
+__version__ = '2.3.0'
+__license__ = 'GPLv3'
+
+__author__ = 'José Antonio Perdiguero López'
+__email__ = 'perdy.hh@gmail.com'
+
+__url__ = 'https://github.com/PeRDy/health-check'
+__description__ = 'Health Check is an application that provides an API to check the health of some parts and some ' \
+                  'utilities like ping requests. This application can works as standalone or included in a Django ' \
+                  'project. '
+
+default_app_config = 'health_check.apps.HealthCheck'