+++ health_check/providers/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from health_check.providers.base import *  # noqa