+++ health_check/static/js/components/HealthCheckApp.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import HealthApp from './HealthApp.jsx';
+import StatsApp from './StatsApp.jsx';
+
+/**
+ * Retrieve the current TODO data from the StatusStore
+ */
+class HealthCheckApp extends React.Component {
+  render() {
+    return (
+      <div>
+        <HealthApp/>
+        <StatsApp/>
+      </div>
+    )
+  }
+}
+
+HealthCheckApp.propTypes = {};
+
+HealthCheckApp.defaultProps = {};
+
+export default HealthCheckApp;