+++ status/providers/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from status.providers.base import *  # noqa